Fix crash issue when dismiss MMI dialog
Dialing an MMI code will start MMIDialogActivity to show MMI
running dialog. After MMI code is completed, the activity will receive
MMI completed event and dismiss the MMI running dialog. But before
receiving this event, the MMIDialogActivity is already finished. So if
we dismiss the dialog in this case, illegalArgumentException will
happen and phone process crashes.

Change-Id: Icfcb6473e202e2e851fa28d389a8a9a41dd2e396